Первый вариант модуляции с ошибками
